initial_seperate_notifications
