+++ trustpoint/home/filters.py
@@ -7,7 +7,7 @@ import django_filters
 from django.db.models import QuerySet
 from django.utils import timezone
 
-from notifications.models import NotificationModel
+from management.models import NotificationModel
 
 
 class NotificationFilter(django_filters.FilterSet):
