+++ trustpoint/home/tests/test_filters.py
@@ -4,7 +4,7 @@ from unittest.mock import Mock, patch
 
 from django.test import TestCase
 from django.utils import timezone
-from notifications.models import NotificationModel
+from management.models import NotificationModel
 
 from ..filters import NotificationFilter
