+++ trustpoint/management/management/commands/check_domain_issuing_ca.py
@@ -6,7 +6,7 @@ from typing import Any
 
 from django.core.management.base import BaseCommand
 from django.utils import timezone
-from notifications.models import NotificationModel, NotificationStatus
+from management.models import NotificationModel, NotificationStatus
 from pki.models import DomainModel
