+++ trustpoint/management/management/commands/check_for_security_vulnerabilities.py
@@ -6,7 +6,7 @@ from typing import Any
 
 from django.core.management.base import BaseCommand
 from django.utils import timezone
-from notifications.models import NotificationModel, NotificationStatus
+from management.models import NotificationModel, NotificationStatus
 
 
 class Command(BaseCommand):
