+++ trustpoint/management/models/pkcs11.py
@@ -1,11 +1,11 @@
-"""Models for the Trustpoint Management app."""
+"""PKCS#11 Token Model."""
 from __future__ import annotations
 
 import hashlib
 import os
 import secrets
 from pathlib import Path
-from typing import TYPE_CHECKING, Any, ClassVar, NoReturn, cast
+from typing import TYPE_CHECKING, Any, NoReturn, cast
 
 import pkcs11  # type: ignore[import-untyped]
 from argon2.low_level import Type, hash_secret_raw
@@ -16,320 +16,10 @@ from django.core.exceptions import ImproperlyConfigured, ValidationError
 from django.db import models
 from django.utils.translation import gettext_lazy as _
 
-from notifications.models import WeakECCCurve, WeakSignatureAlgorithm
-from pki.util.keys import AutoGenPkiKeyAlgorithm
 from trustpoint.logger import LoggerMixin
 
 if TYPE_CHECKING:
     from management.pkcs11_util import Pkcs11AESKey
-
-
-
-class SecurityConfig(models.Model):
-    """Security Configuration model."""
-
-    class SecurityModeChoices(models.TextChoices):
-        """Types of security modes."""
-
-        DEV = '0', _('Testing env')
-        LOW = '1', _('Basic')
-        MEDIUM = '2', _('Medium')
-        HIGH = '3', _('High')
-        HIGHEST = '4', _('Highest')
-
-    security_mode = models.CharField(max_length=6, choices=SecurityModeChoices, default=SecurityModeChoices.LOW)
-
-    auto_gen_pki = models.BooleanField(default=False)
-    auto_gen_pki_key_algorithm = models.CharField(
-        max_length=24, choices=AutoGenPkiKeyAlgorithm, default=AutoGenPkiKeyAlgorithm.RSA2048
-    )
-
-    NOTIFICATION_CONFIGURATIONS: ClassVar[dict[str, dict[str, Any]]] = {
-        SecurityModeChoices.DEV: {
-            'cert_expiry_warning_days': 10,
-            'issuing_ca_expiry_warning_days': 10,
-            'rsa_minimum_key_size': 1024,
-            'weak_ecc_curves': [],
-            'weak_signature_algorithms': [],
-        },
-        SecurityModeChoices.LOW: {
-            'cert_expiry_warning_days': 15,
-            'issuing_ca_expiry_warning_days': 15,
-            'rsa_minimum_key_size': 1024,
-            'weak_ecc_curves': [],
-            'weak_signature_algorithms': [],
-        },
-        SecurityModeChoices.MEDIUM: {
-            'cert_expiry_warning_days': 20,
-            'issuing_ca_expiry_warning_days': 20,
-            'rsa_minimum_key_size': 2048,
-            'weak_ecc_curves': [
-            ],
-            'weak_signature_algorithms': [
-                WeakSignatureAlgorithm.SignatureChoices.MD5,
-                WeakSignatureAlgorithm.SignatureChoices.SHA1,
-            ],
-        },
-        SecurityModeChoices.HIGH: {
-            'cert_expiry_warning_days': 25,
-            'issuing_ca_expiry_warning_days': 25,
-            'rsa_minimum_key_size': 3072,
-            'weak_ecc_curves': [
-            ],
-            'weak_signature_algorithms': [
-                WeakSignatureAlgorithm.SignatureChoices.MD5,
-                WeakSignatureAlgorithm.SignatureChoices.SHA1,
-                WeakSignatureAlgorithm.SignatureChoices.SHA224,
-            ],
-        },
-        SecurityModeChoices.HIGHEST: {
-            'cert_expiry_warning_days': 30,
-            'issuing_ca_expiry_warning_days': 30,
-            'rsa_minimum_key_size': 4096,
-            'weak_ecc_curves': [
-                WeakECCCurve.ECCCurveChoices.SECP160R1,
-                WeakECCCurve.ECCCurveChoices.SECP192R1,
-                WeakECCCurve.ECCCurveChoices.SECP224R1,
-            ],
-            'weak_signature_algorithms': [
-                WeakSignatureAlgorithm.SignatureChoices.MD5,
-                WeakSignatureAlgorithm.SignatureChoices.SHA1,
-                WeakSignatureAlgorithm.SignatureChoices.SHA224,
-            ],
-        },
-    }
-
-    notification_config = models.OneToOneField(
-        'notifications.NotificationConfig',
-        on_delete=models.CASCADE,
-        related_name='security_config',
-        null=True,
-        blank=False,
-        help_text=_('Notification configuration associated with this security level.'),
-    )
-
-    def __str__(self) -> str:
-        """Output as string."""
-        return f'{self.security_mode}'
-
-    def apply_security_settings(self) -> None:
-        """Apply appropriate configuration values based on the security mode."""
-        if self.security_mode and self.notification_config:
-            # Get the default configuration for the selected security level
-            config_values = self.NOTIFICATION_CONFIGURATIONS.get(self.security_mode, {})
-
-            # Apply values to the NotificationConfig
-            self.notification_config.cert_expiry_warning_days = config_values.get(
-                'cert_expiry_warning_days', self.notification_config.cert_expiry_warning_days
-            )
-            self.notification_config.issuing_ca_expiry_warning_days = config_values.get(
-                'issuing_ca_expiry_warning_days', self.notification_config.issuing_ca_expiry_warning_days
-            )
-            self.notification_config.rsa_minimum_key_size = config_values.get(
-                'rsa_minimum_key_size', self.notification_config.rsa_minimum_key_size
-            )
-
-            # Update WeakECCCurve and WeakSignatureAlgorithm relationships
-            weak_ecc_curve_oids = config_values.get('weak_ecc_curves', [])
-            weak_signature_algorithm_oids = config_values.get('weak_signature_algorithms', [])
-
-            weak_ecc_curves = WeakECCCurve.objects.filter(oid__in=weak_ecc_curve_oids)
-            weak_signature_algorithms = WeakSignatureAlgorithm.objects.filter(oid__in=weak_signature_algorithm_oids)
-
-            self.notification_config.weak_ecc_curves.set(weak_ecc_curves)
-            self.notification_config.weak_signature_algorithms.set(weak_signature_algorithms)
-
-            self.notification_config.save()
-
-
-class TlsSettings(models.Model):
-    """TLS settings model."""
-
-    ipv4_address = models.GenericIPAddressField(protocol='IPv4', null=True, blank=True)
-
-    def __str__(self) -> str:
-        """Return a string representation of the TLS settings."""
-        return f"TLS Settings (IPv4: {self.ipv4_address or 'None'})"
-
-    @classmethod
-    def get_first_ipv4_address(cls) -> str:
-        """Get the first IPv4 address or a default value."""
-        try:
-            network_settings = cls.objects.get(id=1)
-            ipv4_address = network_settings.ipv4_address
-            if ipv4_address is None:
-                return '127.0.0.1'
-        except cls.DoesNotExist:
-            return '127.0.0.1'
-        else:
-            return ipv4_address
-
-class AppVersion(models.Model):
-    """Model representing the application version and its last update timestamp."""
-    objects: models.Manager[AppVersion]
-
-    version = models.CharField(max_length=17)
-    container_id = models.CharField(
-        max_length=64,
-        blank=True,
-        default='',
-        help_text=_('Container build ID or hash')
-    )
-    last_updated = models.DateTimeField(auto_now=True)
-
-    class Meta:
-        """Meta options for the AppVersion model."""
-        verbose_name = 'App Version'
-
-    def __str__(self) -> str:
-        """Return a string representation for the AppVersion."""
-        build_info = f' (Build: {self.container_id})' if self.container_id else ''
-        return f'{self.version}{build_info} @ {self.last_updated.isoformat()}'
-
-
-class BackupOptions(models.Model):
-    """A singleton model (we always operate with pk=1) for backup settings.
-
-    We store host/port/user/local_storage, plus either a password or an SSH key.
-    """
-    class AuthMethod(models.TextChoices):
-        """Authentication methods for backup options."""
-        PASSWORD = 'password', 'Password'
-        SSH_KEY   = 'ssh_key',  'SSH Key'
-
-    enable_sftp_storage = models.BooleanField(default=False, verbose_name=_('Use SFTP storage'))
-
-    host = models.CharField(max_length=255, verbose_name=_('Host'), blank=True)
-    port = models.PositiveIntegerField(default=2222, verbose_name=_('Port'), blank=True)
-    user = models.CharField(max_length=128, verbose_name=_('Username'), blank=True)
-
-    auth_method = models.CharField(
-        max_length=10,
-        choices=AuthMethod.choices,
-        default=AuthMethod.PASSWORD,
-        verbose_name=_('Authentication Method')
-    )
-
-    # TODO (Dome): Storing passwords in plain text  # noqa: FIX002
-    password = models.CharField(
-        max_length=128,
-        blank=True,
-        verbose_name=_('Password'),
-        help_text=_('Plain-text password for SFTP.')
-    )
-
-    private_key = models.TextField(
-        blank=True,
-        verbose_name=_('SSH Private Key (PEM format)'),
-        help_text=_('Paste the private key here (PEM).')
-    )
-
-    key_passphrase = models.CharField(
-        max_length=128,
-        blank=True,
-        verbose_name=_('Key Passphrase'),
-        help_text=_('Passphrase for the private key, if any.')
-    )
-
-    remote_directory = models.CharField(
-        max_length=512,
-        blank=True,
-        default='/upload/trustpoint/',
-        verbose_name=_('Remote Directory'),
-        help_text=_('Remote directory (e.g. /backups/) where files should be uploaded. '
-                  'Trailing slash is optional.'),
-    )
-
-    class Meta:
-        """Meta options for the BackupOptions model."""
-        verbose_name = 'Backup Option'
-
-    def __str__(self) -> str:
-        """Return a string representation of the backup options."""
-        return f'{self.user}@{self.host}:{self.port} ({self.auth_method})'
-
-    def save(self, *args: Any, **kwargs: Any) -> None:
-        """Ensure only one instance exists (singleton pattern)."""
-        self.full_clean()
-
-        super().save(*args, **kwargs)
-
-    def clean(self) -> None:
-        """Prevent the creation of more than one instance."""
-        if BackupOptions.objects.exists() and not self.pk:
-            msg = 'Only one BackupOptions instance is allowed.'
-            raise ValidationError(msg)
-
-        return super().clean()
-
-class KeyStorageConfig(models.Model):
-    """Configuration model for cryptographic material storage."""
-
-    class StorageType(models.TextChoices):
-        """Types of cryptographic storage."""
-        SOFTWARE = 'software', _('Software (No Encryption)')
-        SOFTHSM = 'softhsm', _('SoftHSM Container')
-        PHYSICAL_HSM = 'physical_hsm', _('Physical HSM')
-
-    storage_type = models.CharField(
-        max_length=12,
-        choices=StorageType.choices,
-        default=StorageType.SOFTWARE,
-        verbose_name=_('Storage Type'),
-        help_text=_('Type of storage for cryptographic material')
-    )
-
-    hsm_config = models.OneToOneField(
-        'PKCS11Token',
-        on_delete=models.SET_NULL,
-        null=True,
-        blank=True,
-        related_name='crypto_storage_config',
-        verbose_name=_('HSM Configuration'),
-        help_text=_('Associated HSM token configuration (SoftHSM or Physical HSM)')
-    )
-
-    last_updated = models.DateTimeField(
-        auto_now=True,
-        verbose_name=_('Last Updated')
-    )
-
-    class Meta:
-        """Meta options for the KeyStorageConfig model."""
-        verbose_name = _('Crypto Storage Configuration')
-        verbose_name_plural = _('Crypto Storage Configurations')
-
-    def __str__(self) -> str:
-        """Return a string representation of the storage configuration."""
-        status = 'Active' if self.hsm_config is not None else 'Inactive'
-        return f'{self.get_storage_type_display()} ({status})'
-
-    @classmethod
-    def get_config(cls) -> KeyStorageConfig:
-        """Get the crypto storage configuration (singleton).
-
-        Returns:
-            KeyStorageConfig: The configuration instance
-
-        Raises:
-            cls.DoesNotExist: If no configuration exists
-        """
-        return cls.objects.get(pk=1)
-
-    @classmethod
-    def get_or_create_default(cls) -> KeyStorageConfig:
-        """Get the configuration or create a default one.
-
-        Returns:
-            KeyStorageConfig: The configuration instance
-        """
-        config, _ = cls.objects.get_or_create(
-            pk=1,
-            defaults={
-                'storage_type': cls.StorageType.SOFTWARE
-            }
-        )
-        return config
 
 
 class PKCS11Token(models.Model, LoggerMixin):
@@ -1276,23 +966,5 @@ class PKCS11Token(models.Model, LoggerMixin):
             msg = f'Failed to decrypt DEK with BEK: {e}'
             raise ValueError(msg) from e
 
-class LoggingConfig(models.Model):
-    """Logging Configuration model."""
 
-    class LogLevelChoices(models.TextChoices):
-        """Types of log levels."""
-
-        DEBUG = '0', _('Debug')
-        INFO = '1', _('Info')
-        WARNING = '2', _('Warning')
-        ERROR = '3', _('Error')
-        CRITICAL = '4', _('Critical')
-
-    log_level = models.CharField(max_length=8, choices=LogLevelChoices, default=LogLevelChoices.INFO)
-
-    last_updated = models.DateTimeField(auto_now=True)
-
-    def __str__(self) -> str:
-        """Output as string."""
-        return f'{self.log_level}'
