--- trustpoint/notifications/__init__.py
@@ -1,1 +1,0 @@
-"""Module that handles Trustpoint notifications components."""
