--- trustpoint/notifications/admin.py
@@ -1,1 +1,0 @@
-"""Registers models with the Django admin site to manage the application's data."""
