--- trustpoint/notifications/management/commands/__init__.py
@@ -1,1 +1,0 @@
-"""Management commands for notifications."""
