--- trustpoint/notifications/tests/__init__.py
@@ -1,1 +1,0 @@
-"""Test module for the Notifications app."""
