+++ trustpoint/templates/management/settings.html
@@ -218,5 +218,5 @@
       </div>
     </div>
   </div>
-</div>
+  {% include 'management/includes/notification_configuration.html' %}</div>
 {% endblock content %}
