+++ trustpoint/trustpoint/settings.py
@@ -265,7 +265,6 @@ INSTALLED_APPS = [
     'signer.apps.SignerConfig',
     'aoki.apps.AokiConfig',
     'management.apps.ManagementConfig',
-    'notifications.apps.NotificationsConfig',
     'trustpoint_core',
     'django.contrib.admin',
     'django.contrib.auth',
