+++ trustpoint/trustpoint/urls.py
@@ -75,7 +75,6 @@ urlpatterns += [
     path('home/', include('home.urls')),
     path('devices/', include('devices.urls')),
     path('management/', include('management.urls')),
-    path('notifications/', include('notifications.urls')),
     path('i18n/', include('django.conf.urls.i18n')),
     path(
         'jsi18n/',
